% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cc}
\alias{cc}
\title{A markov sampler using Rcpp}
\usage{
cc(sigma, xx, N)
}
\arguments{
\item{sigma}{the variance}

\item{xx}{the Initial value}

\item{N}{Number of random numbers}
}
\value{
Returns random sequence and rejection probability
}
\description{
A markov sampler using Rcpp
}
\examples{
\dontrun{
cc(5,6,100)
}
}
