% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm11.R
\name{gmm11}
\alias{gmm11}
\title{Grey prediction model gmm11 using R}
\usage{
gmm11(x)
}
\arguments{
\item{x}{raw data}
}
\value{
predicted values of raw data
}
\description{
Grey prediction model gmm11 using R
}
\examples{
\dontrun{
x<-runif(100)
z<-gmm11(x)
z
}
}
