% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_kmeans.R
\name{My_kmeans}
\alias{My_kmeans}
\title{Calculate the center of K-mean cluster using R}
\usage{
My_kmeans(data, k, max.iter = 10)
}
\arguments{
\item{data}{a data matrix or a dist object}

\item{k}{the number of cluster}

\item{max.iter}{Maximum number of iterations for clustering. The default value is 10.}
}
\value{
Cluster center for each class
}
\description{
Calculate the center of K-mean cluster using R
}
\examples{
\dontrun{
data<-iris[,1:2]
k<-3
max.iter<-10
z<-My_kmeans(data,k,max.iter)
z
}
}
