\name{SC19001-package}
\alias{SC19001-package}
\alias{SC19001}
\docType{package}
\title{
  Grey prediction model gmm11 using R
}
\description{
  Gray prediction based on existing data。The gray prediction method is to find the regularity of system changes through association analysis and data processing, and generate a strong regular data sequence. Then predict future trends based on it.
}
\author{
Chen Cai

Maintainer: Chen Cai <ustccc@mail.ustc.edu.cn>
}
\keyword{ package }